requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
protobuf==6.33.1
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
scipy==1.13.1
sentry-sdk==2.46.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tomli==2.3.0
torch==2.8.0
torchvision==0.23.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.1
zipp==3.23.0
